修改月報表設定容量欄位的小數點與新增累積量查詢
1.月報表設定(裝置容量與購售裝置容量)修改為Decimal，資料庫需要修改類型(測試完成)
2.累積電量查詢 Bar圖顯示(測試完成)

+++ MICReportSystem/Mysql_Module/ReportConfig.cs
@@ -51,11 +51,11 @@ namespace MICReportSystem.Mysql_Module
         /// <summary>
         /// 裝置容量
         /// </summary>
-        public int DeviceCapacity { get; set; }
+        public decimal DeviceCapacity { get; set; }
         /// <summary>
         /// 購售電容量
         /// </summary>
-        public int PurchaseAndSaleCapacity { get; set; }
+        public decimal PurchaseAndSaleCapacity { get; set; }
         /// <summary>
         /// 比值
         /// </summary>
